requirements.txt
phidata
groq
duckduckgo-search
python-dotenv
phi

